# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ReorderBody {
  ids: number[];
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  isProtected?: boolean;
  createdAt?: string;
}

export interface CreateEmployeeBody {
  name: string;
  role: string;
}

export interface CustomerStatus {
  id: number;
  name: string;
  color: string;
  sortOrder: number;
  isSystem?: boolean;
}

export interface CreateCustomerStatusBody {
  name: string;
  color: string;
  sortOrder?: number;
}

export interface CustomerSource {
  id: number;
  name: string;
  description?: string;
  icon?: string;
}

export interface CreateCustomerSourceBody {
  name: string;
  description?: string;
  icon?: string;
}

export interface ProductType {
  id: number;
  name: string;
  slug?: string;
}

export interface CreateProductTypeBody {
  name: string;
  slug?: string;
}

export interface SupplySource {
  id: number;
  name: string;
  inventoryValue?: number;
}

export interface CreateSupplySourceBody {
  name: string;
  inventoryValue?: number;
}

export interface Product {
  id: number;
  name: string;
  productTypeId?: number;
  productTypeName?: string;
  supplySourceId?: number;
  supplySourceName?: string;
  quantity: number;
  costPrice: number;
  sellPrice: number;
  warrantyMonths?: number;
  note?: string;
  createdAt?: string;
}

export interface CreateProductBody {
  name: string;
  productTypeId?: number;
  supplySourceId?: number;
  quantity: number;
  costPrice: number;
  sellPrice: number;
  warrantyMonths?: number;
  note?: string;
}

export interface Order {
  id: number;
  customerId: number;
  closedAt?: string | null;
  orderCode?: string | null;
  productTypeId?: number | null;
  productTypeName?: string | null;
  supplySourceId?: number | null;
  supplySourceName?: string | null;
  productId?: number | null;
  productName?: string | null;
  customProductName?: string | null;
  revenue: number;
  profit: number;
  warrantyMonths?: number | null;
  warrantyExpiry?: string | null;
  note?: string | null;
  createdAt?: string;
}

export interface Customer {
  id: number;
  code?: string;
  createdAt?: string;
  name: string;
  phone: string;
  address?: string;
  note?: string;
  statusId: number;
  statusName?: string;
  statusColor?: string;
  employeeId: number;
  employeeName?: string;
  sourceId?: number;
  sourceName?: string;
  lastContactAt?: string | null;
  nextContactAt?: string | null;
  orders?: Order[];
  totalRevenue?: number;
  totalProfit?: number;
}

export interface CustomerListResponse {
  customers: Customer[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface CreateOrderBody {
  closedAt?: string | null;
  orderCode?: string | null;
  productTypeId?: number | null;
  supplySourceId?: number | null;
  productId?: number | null;
  customProductName?: string | null;
  revenue: number;
  profit: number;
  warrantyMonths?: number | null;
  note?: string | null;
}

export interface CreateCustomerBody {
  createdAt?: string | null;
  name: string;
  phone: string;
  address?: string | null;
  note?: string | null;
  statusId: number;
  employeeId: number;
  sourceId?: number | null;
  lastContactAt?: string | null;
  nextContactAt?: string | null;
  orders?: CreateOrderBody[];
}

export interface ReportSummary {
  safeProfit: number;
  warrantyRiskAmount: number;
  inventoryCapital: number;
  productsNeedPush?: number;
  inventoryTurnover?: number;
  averageInventoryAge?: number;
  totalCustomers: number;
  totalRevenue: number;
  todayCustomers?: number;
  totalProfit: number;
}

export interface StatusDistributionItem {
  statusId: number;
  statusName: string;
  statusColor?: string;
  count: number;
}

export type EmployeeCustomerStatsStatusBreakdown = { [key: string]: number };

export interface EmployeeCustomerStats {
  employeeId: number;
  employeeName: string;
  statusBreakdown?: EmployeeCustomerStatsStatusBreakdown;
}

export interface DailyTrendItem {
  date: string;
  newCustomers: number;
  closedOrders: number;
}

export interface SourceCustomerStats {
  sourceName: string;
  closed: number;
  notClosed: number;
}

export interface TopSalesItem {
  employeeName: string;
  closeRate: number;
  closedCount: number;
}

export interface MonthlyTrendItem {
  month: string;
  newCustomers: number;
  closedOrders: number;
}

export interface MonthlyProfitItem {
  month: string;
  profit: number;
}

export interface TopCustomerItem {
  customerId: number;
  customerName: string;
  totalRevenue: number;
  totalProfit: number;
}

export interface RevenueBySourceItem {
  sourceName: string;
  totalRevenue: number;
}

export interface RevenueByEmployeeItem {
  employeeName: string;
  totalRevenue: number;
}

export interface WarrantyItem {
  orderId: number;
  orderCode?: string | null;
  customerId: number;
  customerName: string;
  phone: string;
  productTypeName?: string | null;
  supplySourceName?: string | null;
  warrantyExpiry?: string | null;
}

export interface TodayStats {
  todayRevenue: number;
  todayOrders: number;
  todayProfit: number;
}

export interface SalesTrendItem {
  date: string;
  revenue: number;
}

export interface InventorySummary {
  totalCapital: number;
  above30Days: number;
  above60Days: number;
  above90Days: number;
  lowStock: number;
}

export type ListCustomersParams = {
  page?: number;
  pageSize?: number;
  statusId?: number;
  employeeId?: number;
  needFollowUp?: boolean;
  search?: string;
};

export type GetReportSummaryParams = {
  fromDate?: string;
  toDate?: string;
};

export type GetDailyTrendParams = {
  fromDate?: string;
  toDate?: string;
};

export type GetTopCustomersParams = {
  limit?: number;
};

export type GetExpiringWarrantiesParams = {
  days?: number;
};