# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM Customer Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /product-types:
    get:
      operationId: listProductTypes
      tags: [settings]
      summary: List product types
      responses:
        "200":
          description: List of product types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductType"
    post:
      operationId: createProductType
      tags: [settings]
      summary: Create product type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductTypeBody"
      responses:
        "201":
          description: Product type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductType"
  /product-types/{id}:
    put:
      operationId: updateProductType
      tags: [settings]
      summary: Update product type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductTypeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductType"
    delete:
      operationId: deleteProductType
      tags: [settings]
      summary: Delete product type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # SUPPLY SOURCES
  # ========================
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products in inventory
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # CUSTOMERS
  # ========================
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    ProductType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
      required: [id, name]
    CreateProductTypeBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required: [name]

    # ========================
    # SUPPLY SOURCE
    # ========================
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        productTypeId:
          type: integer
        productTypeName:
          type: string
        supplySourceId:
          type: integer
        supplySourceName:
          type: string
        quantity:
          type: integer
        costPrice:
          type: integer
        sellPrice:
          type: integer
        warrantyMonths:
          type: integer
        note:
          type: string
        createdAt:
          type: string
      required: [id, name, quantity, costPrice, sellPrice]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        productTypeId:
          type: integer
        supplySourceId:
          type: integer
        quantity:
          type: integer
        costPrice:
          type: integer
        sellPrice:
          type: integer
        warrantyMonths:
          type: integer
        note:
          type: string
      required: [name, quantity, costPrice, sellPrice]

    # ========================
    # CUSTOMER
    # ========================
...