# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  productTypeId?: number;
  supplySourceId?: number;
  quantity: number;
  costPrice: number;
  sellPrice: number;
  warrantyMonths?: number;
  note?: string;
}