# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  productTypeId?: number;
  productTypeName?: string;
  supplySourceId?: number;
  supplySourceName?: string;
  quantity: number;
  costPrice: number;
  sellPrice: number;
  warrantyMonths?: number;
  note?: string;
  createdAt?: string;
}